IT-2026-09-12-3007: C callback arguments -- pass a Palan function to glibc
Lets a Palan function be passed where a cinclude'd C function expects a
function-pointer callback (qsort's comparator, atexit's handler, etc).
Previously any function with a `pntr(func(...))` parameter was rejected
wholesale as an unsupported signature.

SA (ingestion boundary, PlnSaInternal.h): normalizeCFuncSig now marks a
callback parameter "_callback-param" instead of contributing to
"_unsupported-sig" when callbackSigRepresentable proves every inner
parameter/return type representable -- an irrepresentable inner
signature still reports "function pointer" as before.

SA (call site, PlnSaExpr.cpp): saCallArgs dispatches a "_callback-param"
slot to new sa_func_ref_arg before the argument reaches
registry_.fromJson (which throws for pntr(func)). Only a bare reference
to a Palan function resolved via findPlnFunc (not findVar) is accepted;
its signature must be exactly typeCompat::Identical to the C callback's
inner signature on every parameter and the return type, with pointer
permission checked in the direction the value actually flows (a `const`
C parameter permits `@T` but not `@!T`; the callback's return flows the
other way). Identical-only, not ImplicitWiden, because glibc calls the
Palan function directly with no Palan-side conversion step to widen
through at the ABI boundary. Two new diagnostics:
E_CallbackArgRequiresFunc, E_CallbackSignatureMismatch.

codegen: new ExprKind::FuncRef / FuncRefExpr lowers to a bare LeaLabel of
the function's own (unmangled) label -- the same shape as a string
literal, no dereference. No new VInstr or x86 instruction: the callback
is invoked by glibc's own compiled code, not by Palan.

qsort/bsearch verified end-to-end (`bin/palan` sorts an array via a
Palan comparator). atexit hits a pre-existing, unrelated linker
limitation (build-mgr links with raw `ld`, no crt startup objects, so
`__dso_handle` is undefined) -- recorded on IT-2026-09-12-3009 for the
build-mgr end-to-end tests, not fixed here.

Prerequisite (separate commit): unified sa_expr_call's and
sa_expr_member_call's argument-analysis loops into one saCallArgs, so
this dispatch only needs to be added in one place. That unification
also fixed an unrelated existing bug where an aliased C call skipped
variadic promotion (`S.printf("%f\n", someFlo32)` printed 0.000000
instead of promoting to flo64).

+++ test/testdata/sa/error_139_c_unsupported_func_param.h
@@ -1,1 +1,1 @@
-void set_cb(int (*cb)(int));
+void set_cb(int (*cb)(long double));
